+++ ChefsChoice/Models/RecipesData.swift
@@ -10,7 +10,7 @@ import Foundation
 struct RecipesData: Codable {
     
     let results: [RecipeData]
-
+    
 }
 
 struct RecipeData: Codable {
@@ -19,7 +19,33 @@ struct RecipeData: Codable {
     let title: String
     let image: String
     let preparationMinutes: Int
-    let cookingMinutes: Int
+    let readyInMinutes: Int
     let veryHealthy: Bool
     let aggregateLikes: Int
+    let servings: Int
+    let analyzedInstructions: [AnalyzedInstructions]
+    
 }
+
+struct AnalyzedInstructions: Codable {
+    
+    let steps: [Steps]
+      
+}
+
+struct Steps: Codable {
+    
+    let number: Int
+    let step: String
+    let ingredients: [Ingredients]
+    
+}
+
+struct Ingredients: Codable {
+    
+    let name: String
+}
+
+
+
+
