+++ ChefsChoice/Modules/DetailModule/StepsModel.swift
@@ -10,12 +10,12 @@ import UIKit
 struct StepsModel {
     let readyInMinutes: Int
     let servings: Int
-    let analyzedInstructions: [AnalyzedInstructions]
+    let analyzedInstructions: [AnalyzedInstruction]
     
     static let steps = StepsModel(readyInMinutes: 45, servings: 6, analyzedInstructions: analiz)
 }
 
-struct AnalyzedInstructions {
+struct AnalyzedInstruction {
     let stepNumber: Int
     let stepInformation: String
     let ingredients: [NamesIngedients]
@@ -27,4 +27,4 @@ struct NamesIngedients {
 
 var names = [NamesIngedients(name: "dgdg dgdfgds waf wsr")]
 
-let analiz = [AnalyzedInstructions(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstructions(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstructions(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstructions(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstructions(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstructions(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstructions(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstructions(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names)]
+let analiz = [AnalyzedInstruction(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstruction(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstruction(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstruction(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstruction(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstruction(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstruction(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names),AnalyzedInstruction(stepNumber: 5, stepInformation: "xfgbdfbdbhdb bdb drb dtbd dbhd dd hdryhh drthryhdrge geg eg eg ese e g ege getgetgth rssrtgrst tsg rsrtgg stg srgrstg rgtsr  g gs r grsggt  sgergr sgrstg rsg rtg  sggs rstg rg sgrtgrt g sg rtgt r", ingredients: names)]
